lab7/python-telegram-ai-bot/requirements.txt
python-telegram-bot
python-dotenv
openai
requests